chore: cleanup PR A — namespace clarity, doc updates, deprecation hardening
Low-risk wins from CLEANUP_ROADMAP.md (sections 1, 3, 4, 5 — partial).

- Section 4: clarify DirectiveRegistry vs MetaRule in docs
- Section 5: tighten session-history shim deprecation wording (v0.9.0 target)
- Section 3: rename prompt_synthesizer.py -> prompt_compactor.py + back-compat shim
- Section 1: docstring clarity on alternate spines (graduation/scoring,
  judgment_decay, rules_distillation) — no deletions

DEFERRED to follow-up PRs:
  PR B — section 1 deletes + section 2 metric merging
  PR C — section 3 rule_synthesizer cleanup
  PR D — section 6 integrations/ namespace rename
