+++ Gradata/src/gradata/enhancements/graduation/rules_distillation.py
@@ -1,4 +1,8 @@
-"""
+"""Rules distillation algorithms outside the current production loop.
+
+The canonical live graduation path does not call this module today. It remains
+available for tests, direct callers, and future wiring.
+
 Rules Distillation — Detect Repeated Patterns Ready for Rule Promotion
 ======================================================================
 Layer 1 Enhancement: imports from patterns/ (memory)
